[fix]ディレクトリ構成の修正
