+++ front/app/_components/activity/ActivityCard.tsx
