+++ front/app/_components/hero/HeroCube.tsx
