+++ front/app/_components/skills/SkillCard.tsx
