+++ front/app/_components/skills/Skills.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect } from 'react';
 import gsap from 'gsap';
 import ScrollTrigger from 'gsap/ScrollTrigger';
-import SkillCard from './cards/SkillCard';
+import SkillCard from './SkillCard';
 import {
   SiJavascript,
   SiTypescript,
