+++ front/app/_components/works/WorkCard.tsx
